All-features-complete. Merged.

+++ afcc_web_app/afcc/shipment/shipment-static/javascript/delete_shipment.js
@@ -1,12 +1,15 @@
 function deleteShipment(shipment_id) {
-    axios.delete('/shipments/' + shipment_id)
-    .then((response) => {
-        console.log(response.data)
-        window.location.href = '/shipments'
+    var result = confirm("Delete this shipment?");
+    if (result) {
+        axios.delete('/shipments/' + shipment_id)
+        .then((response) => {
+            console.log(response.data)
+            window.location.href = '/shipments'
 
-    })
-    .catch((error) => {
-        console.log(error)
-        window.location.href = '/shipments'
-    })
+        })
+        .catch((error) => {
+            console.log(error)
+            window.location.href = '/shipments'
+        })
+    }
 }
